backend/requirements.txt

#  BACKEND (FastAPI)
fastapi==0.135.3
uvicorn==0.44.0
python-multipart==0.0.22
pydantic==2.12.5

# AI / MODEL (TensorFlow)
tensorflow==2.13.0
keras==2.13.1
numpy==1.23.5
pillow==12.1.1


# UTILITIES
python-dotenv==1.2.2
requests==2.33.1

 CHATBOT (Groq)
groq==1.1.2


#  OPTIONAL (Google AI - if used)
google-genai==1.70.0